+++ mobile/app/src/main/java/com/piggybank/renote/ui/rekening/Rekening.kt
@@ -6,5 +6,5 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class Rekening(
     val name: String,
-    val uang: Long,
+    val uang: Int,
 ) : Parcelable
